pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
authors = [
    {name = "MuhutDil", github = "https://github.com/MuhutDil"}
    ]
readme = "README.md"
requires-python = ">=3.10.12"
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "dj-database-url>=3.0.1",
    "django>=5.2.4",
    "django-bootstrap5>=25.1",
    "django-crispy-forms>=2.4",
    "django-filter>=25.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["task_manager"]
